rag-query/requirements.txt
fastapi
uvicorn

# Vector DB
pinecone

# Embeddings
sentence-transformers
numpy

# Local LLM
llama-cpp-python

# Utilities
pydantic
python-dotenv